+++ src/main/java/br/edu/ifms/detran/model/Carro.java
@@ -18,7 +18,7 @@ public class Carro implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	private String marca;
 	private Integer modelo;
@@ -36,7 +36,7 @@ public class Carro implements Serializable {
 
 	
 
-	public Carro(Long id, String nome, String marca, Integer modelo, int placa, String cor, Apolice apolice) {
+	public Carro(Integer id, String nome, String marca, Integer modelo, int placa, String cor, Apolice apolice) {
 		super();
 		this.id = id;
 		this.nome = nome;
@@ -49,13 +49,13 @@ public class Carro implements Serializable {
 
 
 
-public Long getId() {
+public Integer getId() {
 		return id;
 	}
 
 
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
